refactor: 자동완성/최근검색어 컴포넌트 분리
